Move website files into website/ directory
All demo/website content (index.html, four-forces/, climb-performance/,
demo/, public/) is now under website/ to separate it clearly from the
library source. Vite root is set to website/ so the build output still
lands at dist/ unchanged; the library build and deploy workflow are
unaffected.

Adds tasks/0003 to track a future migration to Astro + Starlight.
