+++ website/climb-performance/main.ts
@@ -1,5 +1,5 @@
 import '../demo/shared.css'
 import { renderSidebar } from '../demo/sidebar'
-import '../src/components/ClimbPerformance'
+import '../../src/components/ClimbPerformance'
 
 renderSidebar('climb-performance')
