+++ website/index.html
@@ -35,6 +35,6 @@
         </div>
       </main>
     </div>
-    <script type="module" src="/demo/landing.ts"></script>
+    <script type="module" src="./demo/landing.ts"></script>
   </body>
 </html>
